Bug Fixes.
Remove debugging that prevented qdbus being used.

Aparently `who` output formats the date differently in some
circumstances, so let's look backwards from the end for the
process IDs...

What used to be the "alternative" udev format is now the
standard one, and the old standard one is now the alternative
as it seems this is the only way to see unplug events these
days.